week-3/requirements.txt
numpy>=1.24.3
Pillow>=10.0.0
requests>=2.31.0
google-generativeai>=0.3.0
kagglehub>=0.1.0
streamlit>=1.29.0
python-dotenv>=1.0.0
onnxruntime>=1.16.3
tf2onnx==1.16.1
onnx>=1.15.0
tensorflow>=2.16.1
protobuf==3.20.3
h5py>=3.10.0